Move converter tests into their own file
Splits the 37-test `mod tests` block out of `converter.rs` into
`converter_tests.rs`, loaded via `#[cfg(test)] #[path = ...] mod tests;`.
`converter.rs` drops from 648 to ~300 lines and now reads as pure
logic; test changes stay localized to the dedicated test file.
